Move the manifest code from repo to common
