+++ src/AppInstallerCommonCore/AppInstallerCommonCore.vcxproj
@@ -196,6 +196,11 @@
     <ClInclude Include="Public\winget\ExperimentalFeature.h" />
     <ClInclude Include="Public\winget\ExtensionCatalog.h" />
     <ClInclude Include="Public\winget\LocIndependent.h" />
+    <ClInclude Include="Public\winget\Manifest.h" />
+    <ClInclude Include="Public\winget\ManifestInstaller.h" />
+    <ClInclude Include="Public\winget\ManifestLocalization.h" />
+    <ClInclude Include="Public\winget\ManifestValidation.h" />
+    <ClInclude Include="Public\winget\ManifestYamlParser.h" />
     <ClInclude Include="Public\winget\Settings.h" />
     <ClInclude Include="Public\winget\UserSettings.h" />
     <ClInclude Include="Public\winget\Yaml.h" />
@@ -218,6 +223,10 @@
     <ClCompile Include="HttpStream\HttpLocalCache.cpp" />
     <ClCompile Include="HttpStream\HttpRandomAccessStream.cpp" />
     <ClCompile Include="JsonUtil.cpp" />
+    <ClCompile Include="Manifest\Manifest.cpp" />
+    <ClCompile Include="Manifest\ManifestInstaller.cpp" />
+    <ClCompile Include="Manifest\ManifestValidation.cpp" />
+    <ClCompile Include="Manifest\YamlParser.cpp" />
     <ClCompile Include="MsixInfo.cpp" />
     <ClCompile Include="Runtime.cpp" />
     <ClCompile Include="pch.cpp">
