+++ src/AppInstallerCommonCore/Manifest/ManifestValidation.cpp
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 #include "pch.h"
-#include "ManifestValidation.h"
+#include "winget/ManifestValidation.h"
 
 namespace AppInstaller::Manifest
 {
