+++ src/AppInstallerCommonCore/Public/winget/Manifest.h
