+++ src/AppInstallerCommonCore/Public/winget/ManifestInstaller.h
