+++ src/AppInstallerCommonCore/Public/winget/ManifestLocalization.h
