+++ src/AppInstallerCommonCore/Public/winget/ManifestValidation.h
