+++ src/AppInstallerCommonCore/Public/winget/ManifestYamlParser.h
