+++ src/AppInstallerRepositoryCore/Microsoft/SQLiteIndex.h
@@ -2,12 +2,12 @@
 // Licensed under the MIT License.
 #pragma once
 #include "SQLiteWrapper.h"
-#include "Manifest/Manifest.h"
 #include "Microsoft/Schema/ISQLiteIndex.h"
 #include "Microsoft/Schema/Version.h"
 #include "Public/AppInstallerRepositorySearch.h"
 #include <AppInstallerLanguageUtilities.h>
 #include <AppInstallerVersions.h>
+#include <winget/Manifest.h>
 
 #include <chrono>
 #include <filesystem>
